requirements.txt
opencv-python
fastapi
uvicorn
websockets
jinja2
ultralytics
python-dotenv
